requirements.txt
gensim==3.6.0
numpy==1.15.4
pyemd==0.5.1
scikit-learn==0.20.2
scikits.bootstrap==0.3.3
scipy==1.2.0
sklearn==0.0
